+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to
 
 ## \[Unreleased]
 
-\-
+### Added
+
+- Public `TurnStage`, `Game.turn_stage` and `GameState.turn_stage` (rules
+    already had to rely on the private version anyway).
 
 ## \[0.22.0] - 2023-11-16